requirements.txt
backcall==0.1.0
certifi==2019.9.11
cffi==1.13.2
chardet==3.0.4
cloudpickle==1.2.2
colorlog==4.0.2
Cython==0.29.14
decorator==4.4.1
entrypoints==0.3
future==0.18.2
glfw==1.8.4
gym==0.15.4
idna==2.8
imageio==2.6.1
imageio-ffmpeg==0.3.0
ipdb==0.12.2
ipython==7.9.0
ipython-genutils==0.2.0
lockfile==0.12.2
mccabe==0.6.1
moviepy==1.0.1
mpi4py==3.0.3
mujoco-py==2.0.2.5
numpy==1.17.4
opencv-python==4.1.1.26
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.2.1
proglog==0.1.9
prompt-toolkit==2.0.10
ptyprocess==0.6.0
pybullet==2.5.7
pygame==1.9.6
pyglet==1.3.2
Pygments==2.4.2
PyOpenGL==3.1.0
pyquaternion==0.9.5
requests==2.22.0
scipy==1.3.2
six==1.13.0
tqdm==4.38.0
traitlets==4.3.3
urllib3==1.25.7
wcwidth==0.1.7
